minor: apigw 测试代码拆分
